requirements.txt
Flask~=3.0.3
pillow~=10.3.0
torch~=2.3.1
torchvision~=0.18.1